Customs dwell time repo

+++ src/PortRiskMonitor.Application/Interfaces/ICustomsDwellTimeService.cs
@@ -1,11 +1,11 @@
-using PortRiskMonitor.Application.DTOs.Enums;
+using PortRiskMonitor.Infrastructure.CustomsDwellTime;
 using RiskMonitor.Logic;
 
 namespace PortRiskMonitor.Application.Interfaces;
 
 public interface ICustomsDwellTimeService : IKriScore<double>
 {
-    float GetAverageDwellHours();
+    double GetAverageDwellHours();
     uint GetPendingCount();
     uint GetOverdueCount();
     ICollection<CustomsDwellDto> GetDwellDetails();
